chore: use modal for invite code
